Add hack to workaround Windows BLE bug
Making the assumption that this is a Windows bug. The Windows BLE backend in `bleak` returns "Unknwon" for the name. This is because for some reason Advertisement.LocalName in the advertising data is empty even though Powered Up devices have 0x09 Complete Local Name set in the advertising data.

To work around this, we get the name from the device after it is connected.